utilities/ai_chatbot/backend/requirements.txt
python-dotenv==1.0.1
pandas==2.1.4

Flask==2.2.0
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==2.3.7
Flask-Migrate==3.1.0
Flask-APScheduler==1.12.4
Flask-SQLAlchemy==3.0.2

psycopg2~=2.9.3
requests==2.31
uuid~=1.30

#elasticsearch==9.1.0
elasticsearch==7.17.6

#for openai model
openai==1.109.1

# this is google Gemini LLM
google-generativeai==0.8.5

# library for similarity search
rapidfuzz==3.14.2